==> 1_code/00_xbrl_data.Rmd <==
---
title: "Getting Data from the Web: XBRL"
author: "Matthias Uckert"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: 
  html_document:
    toc: true
    toc_depth: 3
    toc_float:
      collapsed: false
    self_contained: no
---

------------------------------------------------------------------------

# Description

In this session we will extract XBRL elements from the files we downloaded in Day 1 - Session 1.

Throughout this script we will use custom functions to download and transform data. All the script specific functions are stored in this folder: **1_code/00_functions/f-xbrl.R** (As a reference I also put the important functions in the script)

```{r setup, include=FALSE, purl=FALSE}
knitr::opts_knit$set(root.dir = normalizePath(here::here())) 
knitr::opts_chunk$set(root.dir = normalizePath(here::here())) 
```

------------------------------------------------------------------------

# Script Setup

```{r script-setup, message=FALSE, warning=FALSE}
library(tidyverse); library(zip); library(furrr); library(stringi)

source("1_code/functions/f-all.R")
source("1_code/00_download_data.R")
source("1_code/functions/f-xbrl.R")

.workers <- min(availableCores() / 2, 6)
```

# Code

------------------------------------------------------------------------

## Paths

```{r paths}
lst_paths <- list(
  dir_main = "2_output/00_xbrl_data",
  dir_sec = "2_output/00_get_edgar_data/documents/",
  path_sec_data = "2_output/00_get_edgar_data/txt_download.rds",
  dir_tmp = "2_output/00_xbrl_data/tmp",
  dir_elements_raw = "2_output/00_xbrl_data/elements_raw",
  dir_elements_prc = "2_output/00_xbrl_data/elements_prc",
  path_elements0 = "2_output/00_xbrl_data/xbrl_elements0.rds",
  path_elements1 = "2_output/00_xbrl_data/xbrl_elements1.rds",
  path_xbrl_lu = "0_data/xbrl_lookup_table.xlsx",
  dir_cache = "2_output/00_xbrl_data/cache/cache_"
) %>% create_dirs()
```

```{r list-sec-files}
tab_files_sec <- list_files_tab(lst_paths$dir_sec, reg = "zip$") %>%
  filter(startsWith(doc_id, "xxx"))
```

------------------------------------------------------------------------

## Extract XBRL Elements

------------------------------------------------------------------------

### Function: xbrl_extract_elements()

To extract the relevant items from the XBRL tagges xml/xsd files we downloaded, we need a more advanced set of functions. Eventually, we achieve this by using a custom function **xbrl_extract_elements()**, which is heavily based on the XBRL library. We had to implement some changes to the original source coude due to a know error in the package (see: https://stackoverflow.com/questions/53651481/schema-file-does-not-exist-in-xbrl-parse-file). Since we already had to change the source code, we implemented some procedures to make the download and the extraction process more efficient.

<details>
<summary>
We use a custom function: **xbrl_extract_elements()** to extract elements ***(click on arrow to show)***.
</summary>
```{r function: xbrl_extract_elements()}
xbrl_extract_elements <- function(.zip, .dir_tmp, .dir_out) {
  zip::unzip(.zip, exdir = .dir_tmp)
  
  
  files_ <- zip_list(.zip)[["filename"]]
  inst_file_ <- files_[endsWith(files_, ".xml")]
  inst_file_ <- inst_file_[!grepl("_lab|_cal|_pre|_def", inst_file_)]
  
  if (length(inst_file_) != 1) {
    return(NULL)
  }
  
  elements_ <- try(own_xbrlDoAll(file.path(.dir_tmp, inst_file_), .dir_tmp, delete.cached.inst = FALSE), silent = TRUE)
  invisible(file.remove(file.path(.dir_tmp, files_)))
  
  if (inherits(elements_, "try-error")) {
    NULL
  } else {
    write_rds(elements_, file.path(.dir_out, gsub("zip$", "rds", basename(.zip))), compress = "gz")
  }
}

own_xbrlDoAll <- function(file.inst, cache.dir = "xbrl.Cache", prefix.out = NULL,
                          verbose = FALSE, delete.cached.inst = TRUE) {
  xbrl <- own_XBRL()
  xbrl$setVerbose(verbose)
  if (!is.null(cache.dir)) {
    xbrl$setCacheDir(cache.dir)
  }
  xbrl$openInstance(file.inst)
  file_schema <- suppressWarnings(xbrl$getSchemaName())[1]
  xbrl$processSchema(file_schema)
  xbrl$processContexts()
  xbrl$processFacts()
  xbrl$processUnits()
  xbrl$processFootnotes()
  xbrl$closeInstance()
  if (delete.cached.inst && gsub("^(http|https|ftp)://.*$", "\\1", file.inst) %in% c("http", "https", "ftp")) {
    xbrl$deleteCachedInstance()
  }
  xbrl.vars <- xbrl$getResults()
  if (!is.null(prefix.out)) {
    if (verbose) {
      cat("Saving data\n")
    }
    write.csv(xbrl.vars$role, file = paste0(prefix.out, "_roles.csv"))
    write.csv(xbrl.vars$element, file = paste0(prefix.out, "_elements.csv"))
    write.csv(xbrl.vars$label, file = paste0(prefix.out, "_labels.csv"))
    write.csv(xbrl.vars$presentation, file = paste0(prefix.out, "_presentations.csv"))
    write.csv(xbrl.vars$definition, file = paste0(prefix.out, "_definitions.csv"))
    write.csv(xbrl.vars$calculation, file = paste0(prefix.out, "_calculations.csv"))
    write.csv(xbrl.vars$context, file = paste0(prefix.out, "_contexts.csv"))
    write.csv(xbrl.vars$fact, file = paste0(prefix.out, "_facts.csv"))
    write.csv(xbrl.vars$unit, file = paste0(prefix.out, "_units.csv"))
    write.csv(xbrl.vars$footnote, file = paste0(prefix.out, "_footnotes.csv"))
  }
  invisible(xbrl.vars)
}
own_XBRL <- function() {
  self <- list(
    element = NULL, role = NULL, label = NULL,
    presentation = NULL, definition = NULL, calculation = NULL,
    context = NULL, unit = NULL, fact = NULL, footnote = NULL
  )
  cache.dir <- NULL
  discovered.files <- NULL
  doc.inst <- NULL
  dname.inst <- NULL
  verbose <- FALSE
  inst.lnkb <- NULL
  
  fixFileName <- function(dname, file.name) {
    if (!(substr(file.name, 1, 5) %in% c("http:", "https"))) {
      if (substr(file.name, 1, 5) == "../..") {
        file.name <- paste0(
          dirname(dirname(dname)),
          "/", substr(file.name, 7, nchar(file.name))
        )
      }
      else if (substr(file.name, 1, 2) == "..") {
        file.name <- paste0(dirname(dname), "/", substr(
          file.name,
          4, nchar(file.name)
        ))
      }
      else {
        file.name <- paste0(dname, "/", file.name)
      }
    }
    file.name
  }
  
  
  setVerbose <- function(newVerbose) {
    oldVerbose <- verbose
    verbose <<- newVerbose
    oldVerbose
  }
  setCacheDir <- function(new.cache.dir) {
    if (!file.exists(new.cache.dir)) {
      dir.create(new.cache.dir)
    }
    cache.dir <<- new.cache.dir
  }
  fileFromCache <- function(file) {
    if (!(gsub("^(http|https|ftp)://.*$", "\\1", file) %in%
          c("http", "https", "ftp"))) {
      return(file)
    }
    bname <- basename(file)
    cached.file <- paste0(cache.dir, "/", bname)
    
    if (!file.exists(cached.file)) {
      if (verbose) {
        cat("Downloading to cache dir...")
      }
      status <- try(download.file(file, cached.file, quiet = !verbose),
                    silent = TRUE
      )
      if (class(status)[1] == "try-error" || status ==
          1) {
        unlink(cached.file)
        stop(status, "\n")
      }
    }
    else {
      if (verbose) {
        cat("Using file from cache dir...\n")
      }
    }
    cached.file
  }
  openInstance <- function(file.inst) {
    dname.inst <<- dirname(file.inst)
    if (!is.null(cache.dir)) {
      file.inst <- fileFromCache(file.inst)
      inst.lnkb <<- file.inst
    }
    doc.inst <<- XBRL::xbrlParse(file.inst)
  }
  deleteCachedInstance <- function() {
    if (verbose) {
      cat("Deleting the following downloaded and/or cached files...\n")
      print(inst.lnkb)
    }
    unlink(inst.lnkb)
    if (verbose) {
      cat("Done...\n")
    }
  }
  getSchemaName <- function() {
    fixFileName(dname.inst, .Call("xbrlGetSchemaName", doc.inst,
                                  PACKAGE = "XBRL"
    ))
  }
  
  processSchema <- function(file, level = 1) {
    if (verbose) {
      cat("Schema: ", file, "\n")
    }
    if (length(which(discovered.files == file)) > 0) {
      if (verbose) {
        cat("Already discovered. Skipping\n")
      }
      return(NULL)
    }
    
    discovered.files <<- c(discovered.files, file)
    dname <- dirname(file)
    if (level >= 1 && !is.null(cache.dir)) {
      if (verbose) {
        cat("Level:", level, "==>", file, "\n")
      }
      file <- fileFromCache(file)
      if (level == 1) {
        inst.lnkb <<- c(inst.lnkb, file)
      }
    }
    doc <- XBRL::xbrlParse(file)
    if (level == 1) {
      processRoles(doc)
    }
    processElements(doc)
    linkbaseNames <- .Call("xbrlGetLinkbaseNames", doc, PACKAGE = "XBRL")
    importNames <- .Call("xbrlGetImportNames", doc, PACKAGE = "XBRL")
    .Call("xbrlFree", doc, PACKAGE = "XBRL")
    for (linkbaseName in linkbaseNames) {
      linkbaseName <- fixFileName(dname, linkbaseName)
      if (verbose) {
        cat(file, " ==> Linkbase: ", linkbaseName, "\n")
      }
      processLinkbase(linkbaseName, level + 1)
    }
    for (importName in importNames) {
      importName <- fixFileName(dname, importName)
      if (verbose) {
        cat(file, " ==> Schema: ", importName, "\n")
      }
      processSchema(importName, level + 1)
    }
  }
  processRoles <- function(doc) {
    if (verbose) cat("Roles\n")
    self$role <<- rbind(
      self$role, .Call("xbrlProcessRoles", doc, PACKAGE = "XBRL")
      )
  }
  processElements <- function(doc) {
    if (verbose) cat("Elements\n")
    self$element <<- rbind(self$element, .Call("xbrlProcessElements", doc, PACKAGE = "XBRL"))
  }
  processLinkbase <- function(file, level) {
    if (verbose) cat("Linkbase: ", file, "\n")
    if (length(which(discovered.files == file)) > 0) {
      if (verbose) cat("Already discovered. Skipping\n")
      return(NULL)
    }
    discovered.files <<- c(discovered.files, file)
    if (level >= 2 && !is.null(cache.dir)) {
      if (verbose) cat("Level:", level, "==>", file, "\n")
      file <- fileFromCache(file)
      inst.lnkb <<- c(inst.lnkb, file)
    }
    doc <- XBRL::xbrlParse(file)
    if (!processLabels(doc)) {
      if (!processPresentations(doc)) {
        if (!processDefinitions(doc)) {
          processCalculations(doc)
        }
      }
    }
    .Call("xbrlFree", doc, PACKAGE = "XBRL")
  }
  
  processLabels <- function(doc) {
    pre.length <- length(self$label)
    self$label <<- rbind(
      self$label, ans <- .Call("xbrlProcessLabels",doc,PACKAGE = "XBRL")
      )
    if (!is.null(ans)) {
      if (verbose) cat("Labels.\n")
      return(TRUE)
    }
    FALSE
  }
  
  processPresentations <- function(doc) {
    pre.length <- length(self$presentation)
    self$presentation <<- rbind(
      self$presentation, ans <- .Call("xbrlProcessArcs",doc, "presentation",PACKAGE = "XBRL")
      )
    
    if (!is.null(ans)) {
      if (verbose) cat("Presentations.\n")
      return(TRUE)
    }
    FALSE
  }
  processDefinitions <- function(doc) {
    pre.length <- length(self$definition)
    self$definition <<- rbind(
      self$definition, ans <- .Call("xbrlProcessArcs",doc, "definition", PACKAGE = "XBRL")
      )
    if (!is.null(ans)) {
      if (verbose) cat("Definitions.\n")
      return(TRUE)
    }
    FALSE
  }
  processCalculations <- function(doc) {
    pre.length <- length(self$calculation)
    self$calculation <<- rbind(
      self$calculation, ans <- .Call("xbrlProcessArcs", doc, "calculation", PACKAGE = "XBRL")
      )
    if (!is.null(ans)) {
      if (verbose) cat("Calculations.\n")
      return(TRUE)
    }
    FALSE
  }
  processContexts <- function() {
    if (verbose) cat("Contexts\n")
    self$context <<- .Call("xbrlProcessContexts", doc.inst, PACKAGE = "XBRL")
  }
  processFacts <- function() {
    if (verbose) cat("Facts\n")
    self$fact <<- .Call("xbrlProcessFacts", doc.inst, PACKAGE = "XBRL")
  }
  processUnits <- function() {
    if (verbose) cat("Units\n")
    self$unit <<- .Call("xbrlProcessUnits", doc.inst, PACKAGE = "XBRL")
  }
  processFootnotes <- function() {
    if (verbose) cat("Footnotes\n")
    self$footnote <<- .Call("xbrlProcessFootnotes", doc.inst, PACKAGE = "XBRL")
  }
  closeInstance <- function() {
    .Call("xbrlFree", doc.inst, PACKAGE = "XBRL")
    doc.inst <<- NULL
  }
  getResults <- function() {
    self
  }
  list(
    setVerbose = setVerbose, setCacheDir = setCacheDir,
    openInstance = openInstance, deleteCachedInstance = deleteCachedInstance,
    getSchemaName = getSchemaName, processSchema = processSchema,
    processContexts = processContexts, processFacts = processFacts,
    processUnits = processUnits, processFootnotes = processFootnotes,
    closeInstance = closeInstance, getResults = getResults
  )
}
```
</details>

------------------------------------------------------------------------

### Extract

```{r extract-elements, message=FALSE, warning=FALSE}
.prc <- list_files_tab(lst_paths$dir_elements_raw)
tab_files_sec_use <- filter(tab_files_sec, !doc_id %in% .prc$doc_id)

# plan("multisession", workers = .workers)
# future_walk(
#   .x = tab_files_sec_use$path, 
#   .f = ~ xbrl_extract_elements(.x, lst_paths$dir_tmp, lst_paths$dir_elements_raw),
#   .options = furrr_options(seed = TRUE)
#   )
# plan("default")

```

------------------------------------------------------------------------

## Process XBRL Elements

<details>
<summary>
We use a custom function: **xbrl_process_elements()** to extract elements from the XBRL tagged .xml files ***(click on arrow to show)***.
</summary>
```{r function: xbrl_process_elements()}
xbrl_process_elements <- function(.inst) {
  chr_desc_ <- .inst$role[["description"]]
  chr_type_ <- .inst$role[["type"]]
  
  if (length(chr_type_) == 0 | length(chr_desc_) == 0) {
    return(NULL)
  }
  
  names(chr_desc_) <- names(chr_type_) <- paste0(chr_type_, "_", chr_desc_)
  
  get_statements <- function(.type, .desc) {
    tab_role_ <- dplyr::filter(.inst$role, description == .desc)
    
    
    tab_skeleton_ <- .inst$presentation %>%
      dplyr::filter(roleId == tab_role_$roleId) %>%
      dplyr::mutate(row_id = dplyr::row_number()) %>%
      tibble::as_tibble()
    
    tab_element_ <- tab_skeleton_ %>%
      dplyr::left_join(.inst$label, by = c("toElementId" = "elementId")) %>%
      dplyr::left_join(.inst$fact, by = c("toElementId" = "elementId")) %>%
      dplyr::left_join(.inst$context, by = c("contextId")) %>%
      dplyr::filter(labelRole == "http://www.xbrl.org/2003/role/label") %>%
      dplyr::filter(!is.na(fact)) %>%
      dplyr::arrange(row_id, startDate, endDate) %>%
      tibble::as_tibble()
    
    if (.type == "Statement") {
      tab_element_ %>%
        dplyr::select(
          tag = toElementId, line_item = labelString,
          unit = unitId, amount = fact, contextId, startDate, endDate
        )
    } else if (.type %in% c("Disclosure", "Document", "Schedule")) {
      tab_element_ %>%
        dplyr::mutate(fact = remove_html_tags(fact)) %>%
        dplyr::select(
          tag = toElementId, line_item = labelString,
          text = fact, contextId, startDate, endDate
        )
    } else {
      warning(paste0("Type not found: ", .type), call. = FALSE)
    }
  }
  
  test <- purrr::map2(chr_type_, chr_desc_, get_statements)
}
```
</details>

------------------------------------------------------------------------

### Process Elements

```{r process-elements, message=FALSE, warning=FALSE}
tab_files_elements_raw <- list_files_tab(lst_paths$dir_elements_raw)
.prc <- list_files_tab(lst_paths$dir_elements_prc)
tab_files_elements_use <- filter(tab_files_elements_raw, !doc_id %in% .prc$doc_id)

plan("multisession", workers = .workers)
future_walk(
  .x = tab_files_elements_use$path,
  .f = ~ {
    read_rds(.x) %>%
      xbrl_process_elements() %>%
      enframe() %>%
      separate(name, c("type", "name"), sep = "_") %>%
      write_rds(file.path(lst_paths$dir_elements_prc, basename(.x)), compress = "gz")
  },
  .options = furrr_options(seed = TRUE)
)
plan("default")
```

------------------------------------------------------------------------

## Read Elements

Here we read all the elements into a single dataframe which we will use in Day 2, to filter for relevant disclosure and financial statements.

```{r read-element-paths}
tab_files_elements_prc <- list_files_tab(lst_paths$dir_elements_prc) %>%
  separate(doc_id, c("tmp", "symbol", "year"), sep = "_|-", convert = TRUE, remove = FALSE)
```

```{r read-elements}
if (!file.exists(lst_paths$path_elements0)) {
  plan("multisession", workers = .workers)
  tab_elements0 <- future_map_dfr(
    .x = set_names(tab_files_elements_prc$path, tab_files_elements_prc$doc_id),
    .f = ~ mutate(read_rds(.x), value = as.list(value)),
    .options = furrr_options(seed = TRUE),
    .id = "doc_id"
  ) %>%
    separate(
      col = doc_id,
      into = c("tmp", "symbol", "year"),
      sep = "_|-",
      convert = TRUE,
      remove = FALSE
    ) %>%
    select(-tmp)
  plan("default")
  write_rds(tab_elements0, lst_paths$path_elements0, compress = "gz")
} else {
  tab_elements0 <- read_rds(lst_paths$path_elements0)
}

show_table(select(tab_elements0, -value), 10)
```

------------------------------------------------------------------------

## Standardize Names

Even though XBRL is intended to standardize the reporting of financial disclosures, the implementation is still not perfect. Especially the names used for describing a specific element are only partially standradized. To overcome this issue we will use a custom lookup table, that assigns each XBRL name a common and standardized description.

```{r standardize-names-read}
tab_lu <- openxlsx::read.xlsx(lst_paths$path_xbrl_lu)
show_table(filter(tab_lu, desc == "Balance Sheet"), 10)
```

------------------------------------------------------------------------

We will just merge our original data with the lookup values to build an standardized table which we can use in further analysis.

```{r standardize-names}
tab_elements1 <- tab_elements0 %>%
  mutate(
    name_adj = name %>%
      stri_replace_all_regex(., "[[:punct:]]", " ") %>%
      stri_replace_all_regex(., "([[:blank:]]|[[:space:]])+", " ") %>%
      trimws() %>%
      tolower()
    ) %>%
  left_join(tab_lu, by = c("type", "name_adj")) %>%
  filter(!is.na(desc))

write_rds(tab_elements1, lst_paths$path_elements1, compress = "gz")

show_table(select(tab_elements1, -value), 10)
# write.table(tab_elements1, "clipboard", sep="\t", row.names = FALSE)
```

------------------------------------------------------------------------

# Own Function Calls

```{r function-calls}
lsf.str()
```

------------------------------------------------------------------------

# Session Info

```{r session-info}
sessioninfo::session_info()
```
