==> 1_code/02_business_case_html.Rmd <==
---
title: "Business Case"
author: "Matthias Uckert"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: 
  html_document:
    toc: true
    toc_depth: 3
    toc_float:
      collapsed: false
    self_contained: no
---

```{r setup, include=FALSE}
knitr::opts_knit$set(root.dir = normalizePath(here::here())) 
knitr::opts_chunk$set(root.dir = normalizePath(here::here()))
```

```{r include=FALSE}
library(tidyverse)
library(janitor)
library(scales)
library(ggthemes)
library(patchwork)
library(kableExtra)
options(scipen = 999)
source("1_code/functions/f-02_business_case.R", encoding = "UTF-8")
source("1_code/functions/v-02_business_case_default_values.R")
```

In this session we will translate the logic of our business case into R.
The output will be a static HTML report. R will give us more flexibility
and performance advantages in comparison to Excel. The transition might
be challenging at first, but if done accurately, functions in R can be
re-used in many other cases.

In the following we will mimic the functionalities of the Excel Business
case.

# Define Input Functions

To derive our our profitability of our product, we first must define
some input function. Namely, we need a function that defines the
**Variable Costs**, **Product Fixed Costs**, and **Revenue**

## Function: Variable Cost

We define a function for the **Variable Costs**: *fnc_var_cost()* with
the following Input Parameters:

-   **.cv (Variable Cost):** Initial Variable

-   **.v (Volume):** The volume we want to derive the total variable
    cost for.

-   **.vb (Base Volume):** The base (minimum) volume. for ever
    additional x (.vd) volume we get an y (.d) discount on our variable
    cost.

-   **.vc (Cap Volume):** The cap (maximum) volume. After this volume no
    discount is granted anymore.

-   **.vd (Volume Discount):** For ever x (.vd) additional unit over the
    base volume (.vb) we get an y (.d) percent discount.

-   **.d (Discount):** The discount we get for ever x (.vd) additional
    unit over the base volume (.vb)

```{r}
.cv <- 500; .v <- 7500; .vb <- 3000; .vc <- 8000; .vd <- 750; .d <- .05
```

```{r}
fnc_var_disc <- function(.v, .vb, .vc, .vd, .d) {
  vu_ <- dplyr::case_when(
    .v < .vb ~ .vb,
    .v > .vc ~ .vc,
    TRUE ~ .v
  )
  floor((vu_ - .vb) / .vd) * .d
}

for (i in c(1000, 3750, 4000, 8000, 9000)) {
  cat("\nDiscount - Volume:", i, ":", percent(fnc_var_disc(i, .vb, .vc, .vd, .d)))
}

```

```{r}
fnc_var_unit <- function(.cv, .v, .vb, .vc, .vd, .d) {
  disc_ <- fnc_var_disc(.v, .vb, .vc, .vd, .d)
  .cv * (1 - disc_)
}

for (i in c(1000, 3000, 4000, 8000, 9000)) {
  cat("\nUnit Cost - Volume:", i, ":", comma(fnc_var_unit(.cv, i, .vb, .vc, .vd, .d)))
}

```

```{r}
fnc_var_cost <- function(.cv, .v, .vb, .vc, .vd, .d) {
  .v * fnc_var_unit(.cv, .v, .vb, .vc, .vd, .d)
}

for (i in c(1000, 3000, 4000, 8000, 9000)) {
  cat("\nTotal Cost - Volume:", i, ":", comma(fnc_var_cost(.cv, i, .vb, .vc, .vd, .d)))
}
```

For convenience we added two helper functions *fnc_var_disc() and
fnc_var_unit()* that calculate the discount as a function of volume and
the unit cost of our product. Those functions will result in so-called
**Step Functions.** Next we plot this functions with the following
values:

-   **.cv (Variable Cost):** 1

-   **.v (Volume):** Range between 500 and 10,000

-   **.vb (Base Volume):** 2,500

-   **.vc (Cap Volume):** 7,500

-   **.vd (Volume Discount):** 1,000

-   **.d (Discount):** 5.00%

```{r echo=FALSE}
.seq <- seq(500, 10000, 5)
tab_vc <- get_input_table(
  .cv = .def_cv, .v = .seq, .vb = .def_vb, .vc = .def_vc, .vd = .def_vd, .d = .def_d
  )
tab_vc <- tab_vc %>%
  mutate(
    discount = fnc_var_disc(.v = v, .vb = vb, .vc = vc, .vd = vd, .d = d),
    var_unit = fnc_var_unit(.cv = cv, .v = v, .vb = vb, .vc = vc, .vd = vd, .d = d),
    var_cost = fnc_var_cost(.cv = cv, .v = v, .vb = vb, .vc = vc, .vd = vd, .d = d)
  )

gg_vc_unit <- plot_points(tab_vc, v, var_unit, NULL, NULL, "Unit Variable Cost", .size = 11)
gg_vc_tot  <- plot_points(tab_vc, v, var_cost, NULL, NULL, "Total Variable Cost", .size = 11)
gg_vc_unit + gg_vc_tot
```

## Function: Fixed Cost

We define a function for the **Fixed Costs**: *fnc_var_cost()* with the
following Input Parameters:

-   **.cf (Variable Cost):** Initial Fixed Costs

-   **.v (Volume):** The volume we want to derive the total variable
    cost for.

-   **.vm (Volume Multiplier):** Fixed Cost double at this point

```{r}
.cf<- 50000; .vm <- 1000
```

```{r message=FALSE, warning=FALSE}
fnc_fix_mult <- function(.v, .vm) {
  ceiling(.v / .vm)
}

for (i in c(1000, 3000, 4000, 8000, 9000)) {
  cat("\nFixed Cost Multiplier - Volume:", i, ":", comma(fnc_fix_mult(i, .vm)))
}
```

```{r message=FALSE, warning=FALSE}
fnc_fix_cost <- function(.cf, .v, .vm) {
  .cf * fnc_fix_mult(.v, .vm)
}

for (i in c(1000, 3000, 4000, 8000, 9000)) {
  cat("\nFixed Cost - Volume:", i, ":", comma(fnc_fix_cost(.cf, i, .vm)))
}
```

```{r message=FALSE, warning=FALSE}
fnc_fix_unit <- function(.cf, .v, .vm) {
  (.cf * fnc_fix_mult(.v, .vm)) / .v
}

for (i in c(1000, 3000, 4000, 8000, 9000)) {
  cat("\nUnit Fixed Cost - Volume:", i, ":", comma(fnc_fix_unit(.cf, i, .vm)))
}
```

For convenience we added two helper functions *fnc_fix_disc() and
fnc_fix_unit()* that calculate the multiplier as a function of volume
and the directly attributable unit fix cost of our product. Next we plot
this functions with the following values:

-   **.cf (Fixed Cost):** 1

-   **.v (Volume):** Range between 500 and 10,000

-   **.vm (Volume Multiplier):** 1,000

```{r echo=FALSE, message=FALSE, warning=FALSE}
.seq <- seq(500, 10000, 5)
tab_cf <- get_input_table(.cf = .def_cf, .v = .seq, .vm = .def_vm) %>%
  mutate(
    multiple = fnc_fix_mult(.v = v, .vm = vm),
    var_unit = fnc_fix_unit(.cf = cf, .v = v, .vm = vm),
    var_cost = fnc_fix_cost(.cf = cf, .v = v, .vm = vm)
  )

gg_cf_unit <- plot_points(tab_cf, v, var_unit, NULL, NULL, "Unit Fixed Cost", .size = 11)
gg_cf_tot  <- plot_points(tab_cf, v, var_cost, NULL, NULL, "Total Fixed Cost", .size = 11)
gg_cf_unit + gg_cf_tot
```

## Function: Price (Revenue)

We define a function for the **Revenue**: *fnc_rev()* with the following
Input Parameters:

-   **.v (Volume):** The volume we want to derive the total variable
    cost for.

-   **.p (Price):** Initial Price

To do so, we first model the price as a function of volume by using a
so-called sigmoid function.

We limit the resulting S-Curve between .6 to 1. The idea behind this is,
that for every addition unit we want to sell, we have to lower the price
by a certain amount.

```{r}
.p <- 750
```

```{r message=FALSE, warning=FALSE}
fnc_rev_sigmoid <- function(.v) {
  (1 - (1 / (1 + exp(-((.v / 1000) - 4.25))))) * .4 + .6
}

for (i in c(1000, 3000, 4000, 8000, 9000)) {
  cat("\nSigmoid - Volume:", i, ":", percent(fnc_rev_sigmoid(i)))
}
```

```{r message=FALSE, warning=FALSE}
fnc_rev_unit <- function(.p, .v) {
  .p * fnc_rev_sigmoid(.v)
}

for (i in c(1000, 3000, 4000, 8000, 9000)) {
  cat("\nUnit Revenue (Price) - Volume:", i, ":", comma(fnc_rev_unit(.p, i)))
}
```

```{r message=FALSE, warning=FALSE}
fnc_rev <- function(.p, .v) {
  .v * fnc_rev_unit(.p, .v)
}

for (i in c(1000, 3000, 4000, 8000, 9000)) {
  cat("\nRevenue- Volume:", i, ":", comma(fnc_rev(.p, i)))
}
```

Next we plot this functions with the following values:

-   **.p(Price):** 750

-   **.v (Volume):** Range between 500 and 10,000

```{r echo=FALSE, message=FALSE, warning=FALSE}
.seq <- seq(500, 10000, 5)
tab_rv <- get_input_table(.p = .def_p, .v = .seq)
tab_rv <- tab_rv %>%
  mutate(
    sigmoid = fnc_rev_sigmoid(.v = v),
    price    = fnc_rev_unit(.p = p, .v = v),
    revenue  = fnc_rev(.p = p, .v = v)
  )

gg_rv_unit <- plot_points(tab_rv, v, price, NULL, NULL, "Price", .size = 11)
gg_rv_tot  <- plot_points(tab_rv, v, revenue, NULL, NULL, "Revenue", .size = 11)
gg_rv_unit + gg_rv_tot
```

# Define Other Functions

After we defined the input functions for variable cost, fixed cost and
price, we will program functions that give us the the contribution
margins and the profit of our business case. We will use a dataframe
logic which is very similar to Excel, but with significant more
flexibility.

## Function: Input Table

First we define a function with all the input variables, specifically:
[**.v (Volume)**]{.underline}, .[**p (Price)**]{.underline}, [**.cv
(Variable Cost)**]{.underline}, [**.cf (Fixed Cost)**]{.underline},
[**.vb (Base Volume - for Variable Cost)**]{.underline}, [**.cv (Cap
Volume - for Variable Cost)**]{.underline}, [**.vd (Volume Discount -
for Variable Cost)**]{.underline}, [**.d (Discount - for Variable
Cost)**]{.underline}, [**.vm (Volume Multiplier - for Fixed
Cost)**]{.underline}, .[**ca (Company Cost)**.]{.underline}

```{r}
.ca <- 250000
```

```{r}
get_input_table <- function(.v = 0, .p = 0, .cv = 0, .cf = 0, .vb = 0, .vc = 0, 
                            .vd = 0, .d = 0, .vm = 0, .ca = 0) {
  tibble::tibble(
    v = .v, p = .p, cv = .cv, cf = .cf, ca = .ca, vb = .vb, vc = .vc,
    vd = .vd, d = .d, vm = .vm,
  )
}

get_input_table(
  .v = .v, .p = .p, .cv = .cv, .cf = .cf, .vb = .vb, .vc = .vc,
  .vd = .vd, .d = .d, .vm = .vm, .ca = .ca
) %>%
  mutate(
    across(!matches("^d$"), comma), 
    d = percent(d)
    ) %>%
  kbl(align = "c") %>%
  kable_paper(full_width = FALSE)
```

## Function: Profit

Next we define a function that calculates the Contribution Margins and
the Profit for our input factor. This function takes a dataframe as only
argument which is the dataframe we produce with the function
**get_input_table()**

```{r}
get_input_table(
  .v = .v, .p = .p, .cv = .cv, .cf = .cf, .vb = .vb, .vc = .vc,
  .vd = .vd, .d = .d, .vm = .vm, .ca = .ca
) %>%
  get_profit() %>%
  mutate(
    across(!matches("^d$|discount"), comma), 
    d = percent(d),
    discount = percent(discount)
    ) %>%
  select(discount:profit) %>%
  kbl(align = "c") %>%
  kable_paper(full_width = FALSE)
```

## Functions: Other

We also make some functions that will help us to reshape the data.

```{r}
get_cm_table <- function(.tab) {
  .tab %>%
    get_profit() %>%
    dplyr::select(
      Revenue = revenue,
      `Variable Cost` = var_cost,
      `Contribution Margin I` = cm1,
      `Fixed Cost` = fix_cost,
      `Contribution Margin II` = cm2,
      `Company Cost` = ca,
      `Profit` = profit
    ) %>%
    tidyr::pivot_longer(cols = dplyr::everything()) %>%
    dplyr::mutate(
      perc = value / first(value),
      value = scales::comma(value),
      perc = scales::percent(perc)
    )
}
```

```{r}
get_amortization <- function(.tab, .years, .invest, .discount) {
  .tab %>%
    tidyr::expand_grid(year = seq_len(.years)) %>%
    dplyr::select(year, dplyr::everything()) %>%
    get_profit() %>%
    dplyr::mutate(dcf = profit / ((1 + .discount)^year)) %>%
    dplyr::select(year, dcf) %>%
    tibble::add_row(year = 0, dcf = -.invest) %>%
    dplyr::arrange(year) %>%
    dplyr::mutate(csum = cumsum(dcf)) %>%
    dplyr::mutate(
      col1 = dplyr::if_else(csum < 0, "red", "green"),
      col2 = dplyr::if_else(dcf < 0, "red", "green")
    ) %>%
    dplyr::select(year, dcf, csum, col1, col2)
}
```

# CM I, CM II, and Profit

Using the Functions **get_input_table()**, **get_profit()**, and
**get_cm_table()**, we can show a scenario analysis for the Contribution
Margin and Profit.

```{r}
tab_ <- get_input_table(
  .v = .def_v, .p = .def_p, .cv = .def_cv, .cf = .def_cf,
  .vb = .def_vb, .vc = .def_vc, .vd = .def_vd, .d = .def_d,
  .vm = .def_vm, .ca = .def_ca
)

s1 <- mutate(tab_, v = 2500) %>%
  get_profit() %>%
  get_cm_table() %>%
  rename(`# 2,500` = value, `% 2,500` = perc)

s2 <- mutate(tab_, v = 5000) %>%
  get_profit() %>%
  get_cm_table() %>%
  rename(`# 5,000` = value, `% 5,000` = perc)

s3 <- mutate(tab_, v = 7500) %>%
  get_profit() %>%
  get_cm_table() %>%
  rename(`# 7,500` = value, `% 7,500` = perc)

s4 <- mutate(tab_, v = 10000) %>%
  get_profit() %>%
  get_cm_table() %>%
  rename(`# 10,000` = value, `% 10,000` = perc)

s1 %>%
  left_join(s2, by = "name") %>%
  left_join(s3, by = "name") %>%
  left_join(s4, by = "name") %>%
  kableExtra::kbl(align = "lcccccc", col.names = NULL) %>%
  kableExtra::kable_paper(html_font = "Times New Roman") %>%
  kableExtra::column_spec(column = c(1, 3, 5, 7), border_right = TRUE) %>%
  kableExtra::row_spec(row = c(1, 3, 5, 7), bold = TRUE) %>%
  kableExtra::add_header_above(
    c(" " = 1, "Volume: 2,500" = 2, "Volume: 5,000" = 2, "Volume: 7,500" = 2, "Volume: 10,000" = 2)
    )
```

# Sensitivity

Next we look at the sensitivity of the profit to the input factors:
**Volume, Price, Variable Cost, Fixed Cost, and Volume Discount.** We
don't need to program an new function for this. The vectorized functions
we defined up until now are sufficient, the only thing we need to do is
adding a range for the input variable we want to observe to the
**get_input_table()** function.

```{r}
.seq <- seq(500, 10000, 25)
get_input_table(
  .v = .seq, .p = .def_p, .cv = .def_cv, .cf = .def_cf,
  .vb = .def_vb, .vc = .def_vc, .vd = .def_vd, .d = .def_d,
  .vm = .def_vm, .ca = .def_ca
) %>%
  get_profit() %>%
  plot_profits(v, profit, .xlab = "Volume", "Profit", "Profit (Volume)", .size = 11)

.seq <- seq(500, 1000, 5)
get_input_table(
  .v = .def_v, .p = .seq, .cv = .def_cv, .cf = .def_cf,
  .vb = .def_vb, .vc = .def_vc, .vd = .def_vd, .d = .def_d,
  .vm = .def_vm, .ca = .def_ca
) %>%
  get_profit() %>%
  plot_profits(p, profit, .xlab = "Price", "Profit", "Profit (Price)", .size = 11)

.seq <- seq(250, 1000, 5)
get_input_table(
  .v = .def_v, .p = .def_p, .cv = .seq, .cf = .def_cf,
  .vb = .def_vb, .vc = .def_vc, .vd = .def_vd, .d = .def_d,
  .vm = .def_vm, .ca = .def_ca
) %>%
  get_profit() %>%
  plot_profits(cv, profit, .xlab = "Variable Cost", "Profit", "Profit (Variable Cost)", .size = 11)

.seq <- seq(25000, 100000, 1000)
get_input_table(
  .v = .def_v, .p = .def_p, .cv = .def_cv, .cf = .seq,
  .vb = .def_vb, .vc = .def_vc, .vd = .def_vd, .d = .def_d,
  .vm = .def_vm, .ca = .def_ca
) %>%
  get_profit() %>%
  plot_profits(cf, profit, .xlab = "Fixed Cost", "Profit", "Profit (Fixed Cost)", .size = 11)

.seq <- seq(.01, .1, .001)
get_input_table(
  .v = .def_v, .p = .def_p, .cv = .def_cv, .cf = .def_cv,
  .vb = .def_vb, .vc = .def_vc, .vd = .def_vd, .d = .seq,
  .vm = .def_vm, .ca = .def_ca
) %>%
  get_profit() %>%
  plot_profits(d, profit, .xlab = "Discount", "Profit", "Profit (Discount)", .size = 11)
```

# Amortization

With the function get_amortization(), we can easily plot the
amortization time of our business case. Again this function takes the
dataframe of get_input_table() as input. Additionally, it has other 3
Arguments: **.years (NUmber of Years to plot), .invest (Initial
Investment), .discount (Required Rate of Return)**

```{r echo=FALSE}
tab_ <- get_input_table(
  .v = .def_v, .p = .def_p, .cv = .def_cv, .cf = .def_cf,
  .vb = .def_vb, .vc = .def_vc, .vd = .def_vd, .d = .def_d,
  .vm = .def_vm, .ca = .def_ca
) 
tab_ <- get_amortization(tab_, 15, .invest = 1000000, .discount = .08)

plot_amortization(tab_, 11)
```

# Monte Carlo

Up until here, everything we did you could also have done in Excel (and
most of it we have in our Excel Business Case). One aspect we cannot do
in Excel (At least not without VBA, or Add-Ins) is to perform a Monte
Carlo Analysis. Given that we have many input factors, and some of our
inputs are related non-linearly to other inputs, it is a priori not
clear how profitable out business case is under specific assumption. A
Monte Carlo Analysis will help us to evaluate this, by simulating random
inputs (with defined probability distributions) and calculates the
profit. With this we can observe the profit distribution and get a
feeling how robust our business model is to changing input factors
(uncertainty).

```{r echo=FALSE, message=FALSE, warning=FALSE}
set.seed(123)
.n = 1000000
tab_mc <- get_input_table(
  .v = rnorm(n = .n, mean = .def_v, sd = .def_vsd), 
  .p = rnorm(n = .n, mean = .def_p, sd = .def_psd), 
  .cv = rnorm(n = .n, mean = .def_cv, sd = .def_cvsd), 
  .cf = rnorm(n = .n, mean = .def_cf, sd = .def_cfsd),
  .vb = rnorm(n = .n, mean = .def_vb, sd = .def_vbsd), 
  .vc = rnorm(n = .n, mean = .def_vc, sd = .def_vcsd), 
  .vd = rnorm(n = .n, mean = .def_vd, sd = .def_vdsd), 
  .d  = rnorm(n = .n, mean = .def_d, sd = .def_dsd),
  .vm = rnorm(n = .n, mean = .def_vm, sd = .def_vmsd), 
  .ca = rnorm(n = .n, mean = .def_ca, sd = .def_casd)
) %>% get_profit()

plot_monte_carlo(tab_mc, .size = 11)
```

```{r echo=FALSE}
tab_mc %>%
  summarise(
    `Profit > 0` = percent(sum(profit > 0) / n()),
    `Profit < 0` = percent(sum(profit < 0) / n()),
    `Mean` = comma(mean(profit)),
    `Median` = comma(median(profit)),
    `Minimum` = comma(min(profit)),
    `Maximum` = comma(max(profit)),
  ) %>%
  pivot_longer(everything()) %>%
  kableExtra::kbl(align = "lc", col.names = NULL) %>%
  kableExtra::kable_paper(html_font = "Times New Roman") %>%
  kableExtra::kable_styling(full_width = FALSE, position = "center")

```
