==> 1_code/03_business_case_shiny.Rmd <==
---
title: "Business Case"
author: "Matthias Uckert"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: 
  html_document:
    toc: true
    toc_depth: 3
    toc_float:
      collapsed: false
    self_contained: no
---

```{r setup, include=FALSE}
knitr::opts_knit$set(root.dir = normalizePath(here::here())) 
knitr::opts_chunk$set(root.dir = normalizePath(here::here()))
```

Last, we translate our code into a reactive framework to rn a shiny
application. If you have already cloned the repository your can run the
app with the command:

```{r eval=FALSE, message=FALSE, warning=FALSE}
library(shiny)
runApp("1_code/03_shiny_business_case.R")
```

We will look at the dashboard in the class. Below some screenshots to
show the Layout of the dashboard.

# Input Function

![Shiny - Input Functions (Variable
Costs)](images/Shiny%20-%20Input%20Functions%20(Variable%20Costs).png)

![Shiny - Input Functions (Fixed
Costs)](images/Shiny%20-%20Input%20Functions%20(Fixed%20Costs).png)

![Shiny - Input Functions
(Price)](images/Shiny%20-%20Input%20Functions%20(Price).png)

# Contribution Margin

![Shiny - Contribution
Margin](images/Shiny%20-%20Contribution%20Margin.png)

# Sensitivity Analysis

![Shiny - Sensitivity Analysis
(Volume)](images/Shiny%20-%20Sensitivity%20Analysis%20(Volume).png)

![Shiny - Sensitivity Analysis
(Price)](images/Shiny%20-%20Sensitivity%20Analysis%20(Price).png)

![Shiny - Sensitivity Analysis (Variable
Cost)](images/Shiny%20-%20Sensitivity%20Analysis%20(Variable%20Cost).png)

![Shiny - Sensitivity Analysis (Fixed
Cost)](images/Shiny%20-%20Sensitivity%20Analysis%20(Fixed%20Cost).png)

![Shiny - Sensitivity Analysis (Volume
Discount)](images/Shiny%20-%20Sensitivity%20Analysis%20(Volume%20Discount).png)

# Monte Carlo Simulation

![Shiny - Monte Carlo
Analysis](images/Shiny%20-%20Monte%20Carlo%20Analysis.png)

# Amortization

![Shiny - Amortization
Analysis](images/Shiny%20-%20Amortization%20Analysis.png)
